Hotbar Bug Fixes
Previously, since arrays act like lists as their size changes when elements are added/removed, this caused an issue in which the order of array elements did not reflect the order shown in the hotbar.

To resolve this, I padded out the hotbar item array with null values, changed the way null checks are handled when finding duplicate items, and added a "FirstEmptySlotNo" variable to insert new items into the first empty slot found.

Therefore, when a new item gets added to the hotbar, it will check the full array if it has a duplicate item in any of its slots. It will also record the first empty slot if it encounters one. If a duplicate item has been found, it will only increase the quantity of that slot. Otherwise, it will replace the first empty slot with the new item instead of adding a new element.

Additionally, I resolved a validity error on the "Equip" method.