Applied Prompt UI to Pickups & Other Things
Now the pickup items has a handy button prompt when near! The pickup colliders now react to the detector instead of the player for consistency, and to make the IK look less back-breaking.

I believe I did other things, but I cannot remember for now. Maybe improvements on Feed IK and animal detection?